More general Region substitution based on boundary
Until now, it was only possible to change the color of a Region of in the
pattern if its image under the found Morphism was also a Region.
Now the fill operation uses a Boundary (set of Sides) to define the area. So if
the image of a Region is not a Region, it can still be repainted. This is the
case for holes.